Extract SerializedCrdts
