+++ packages/liveblocks-core/src/LiveMap.ts
@@ -3,17 +3,16 @@ import { AbstractCrdt, OpSource } from "./AbstractCrdt";
 import { nn } from "./lib/assert";
 import { freeze } from "./lib/freeze";
 import type {
-  IdTuple,
   LiveMapUpdates,
   LiveNode,
   Lson,
   ParentToChildNodeMap,
-  SerializedMap,
 } from "./types";
-import { CrdtType } from "./types";
 import type { ToImmutable } from "./types/Immutable";
 import type { CreateChildOp, CreateMapOp, Op } from "./types/Op";
 import { OpCode } from "./types/Op";
+import type { IdTuple, SerializedMap } from "./types/SerializedCrdt";
+import { CrdtType } from "./types/SerializedCrdt";
 import {
   creationOpToLiveNode,
   deserialize,
