+++ packages/liveblocks-core/src/LiveRegister.ts
@@ -1,17 +1,12 @@
 import type { ApplyResult, ManagedPool } from "./AbstractCrdt";
 import { AbstractCrdt } from "./AbstractCrdt";
 import { nn } from "./lib/assert";
-import type {
-  IdTuple,
-  Json,
-  LiveNode,
-  ParentToChildNodeMap,
-  SerializedRegister,
-} from "./types";
-import { CrdtType } from "./types";
+import type { Json, LiveNode, ParentToChildNodeMap } from "./types";
 import type { Immutable } from "./types/Immutable";
 import type { CreateChildOp, CreateRegisterOp, Op } from "./types/Op";
 import { OpCode } from "./types/Op";
+import type { IdTuple, SerializedRegister } from "./types/SerializedCrdt";
+import { CrdtType } from "./types/SerializedCrdt";
 
 /**
  * INTERNAL
