+++ packages/liveblocks-core/src/__tests__/LiveObject.test.ts
@@ -1,9 +1,10 @@
 import { LiveList } from "..";
 import { RoomScope } from "../AuthToken";
 import { LiveObject } from "../LiveObject";
-import type { IdTuple, SerializedCrdt } from "../types";
-import { CrdtType, WebsocketCloseCodes } from "../types";
+import { WebsocketCloseCodes } from "../types";
 import { OpCode } from "../types/Op";
+import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
+import { CrdtType } from "../types/SerializedCrdt";
 import {
   createSerializedList,
   createSerializedObject,
